refactor: Split problem.py and callback.py into focused modules
