+++ src/saealib/callback/__init__.py
@@ -1,0 +1,41 @@
+"""
+Callback module.
+
+This module contains event classes and the callback manager for
+the optimization lifecycle.
+"""
+
+from saealib.callback.events import (
+    Event,
+    GenerationEndEvent,
+    GenerationStartEvent,
+    PostAskEvent,
+    PostCrossoverEvent,
+    PostEvaluationEvent,
+    PostMutationEvent,
+    PostSurrogateFitEvent,
+    RunEndEvent,
+    RunStartEvent,
+    SurrogateEndEvent,
+    SurrogateStartEvent,
+)
+from saealib.callback.handlers import logging_generation, logging_generation_hv
+from saealib.callback.manager import CallbackManager
+
+__all__ = [
+    "CallbackManager",
+    "Event",
+    "GenerationEndEvent",
+    "GenerationStartEvent",
+    "PostAskEvent",
+    "PostCrossoverEvent",
+    "PostEvaluationEvent",
+    "PostMutationEvent",
+    "PostSurrogateFitEvent",
+    "RunEndEvent",
+    "RunStartEvent",
+    "SurrogateEndEvent",
+    "SurrogateStartEvent",
+    "logging_generation",
+    "logging_generation_hv",
+]
