+++ tests/test_callback.py
@@ -37,7 +37,8 @@ from saealib.callback import (
 )
 from saealib.context import OptimizationContext
 from saealib.population import Archive, Population, PopulationAttribute
-from saealib.problem import Problem, SingleObjectiveComparator
+from saealib.comparators import SingleObjectiveComparator
+from saealib.problem import Problem
 from saealib.surrogate.manager import (
     EnsembleSurrogateManager,
     GlobalSurrogateManager,
